modified

+++ motorcode/motorcode/main.cpp
@@ -7,28 +7,27 @@ long double map(long double x, long double in_min, long double in_max, long doub
 {
 	return (x - in_min) * (out_max - out_min) / (in_max - in_min) + out_min;
 }
-void ellipticalSquareToDisc(long double x, long double y, long double& u, long double& v)
+void ellipticalSquareToDisc(double x, double y, double& u, double& v)
 {
     u = x * sqrt(1.0 - y*y/2.0);
     v = y * sqrt(1.0 - x*x/2.0);
 }
 
-void ellipticalDiscToSquare(long double u, long double v, long double& x, long double& y)
+void ellipticalDiscToSquare(double u, double v, double& x, double& y)
 {
-    long double u2 = u * u;
-    long double v2 = v * v;
-    long double twosqrt2 = 2.0 * sqrt(2.0);
-    long double subtermx = 2.0 + u2 - v2;
-    long double subtermy = 2.0 - u2 + v2;
-    long double termx1 = subtermx + u * twosqrt2;
-    long double termx2 = subtermx - u * twosqrt2;
-    long double termy1 = subtermy + v * twosqrt2;
-    long double termy2 = subtermy - v * twosqrt2;
+    double u2 = u * u;
+    double v2 = v * v;
+    double twosqrt2 = 2.0 * sqrt(2.0);
+    double subtermx = 2.0 + u2 - v2;
+    double subtermy = 2.0 - u2 + v2;
+    double termx1 = subtermx + u * twosqrt2;
+    double termx2 = subtermx - u * twosqrt2;
+    double termy1 = subtermy + v * twosqrt2;
+    double termy2 = subtermy - v * twosqrt2;
     x = 0.5 * sqrt(termx1) - 0.5 * sqrt(termx2);
     y = 0.5 * sqrt(termy1) - 0.5 * sqrt(termy2);
 
-}
-int main (void)
+}int main (void)
 {
     TCCR0 |= 1<<WGM00 | 1<< WGM01  | 1<< COM01 |1 <<CS00;
 	DDRB |= 1<< PB3 | 1<<PINB0 | 1<<PINB1;
@@ -53,20 +52,20 @@ int main (void)
 		ADMUX ^= (1<<MUX0);
 
 
-		long double x1 = x2 - 512;
-		long double y1 = 512 - y2;
-		x1=map(x1,-512,512,-1,1);
-		y1=map(y1,-512,512,-1,1);
-		long double xans,yans;
-		ellipticalSquareToDisc(x1, y1, xans, yans);
+		double x1 = x2 - 512;
+    double y1 = 512 - y2;
+    x1=map(x1,-512,512,-1,1);
+    y1=map(y1,-512,512,-1,1);
+    double xans,yans;
+    ellipticalSquareToDisc(x1, y1, xans, yans);
 
 
 
-		long double x = (xans * 0.707) + (yans * 0.707);
-		long double y = (-xans * 0.707) + (yans * 0.707);
-		ellipticalDiscToSquare(x,y,xans,yans);
-		int x3=map(xans,-0.991273,0.991273,-255,255);
-		int y3=map(yans,-0.991273,0.991273,-255,255);
+    double x = (xans * 0.707) + (yans * 0.707);
+    double y = (-xans * 0.707) + (yans * 0.707);
+    ellipticalDiscToSquare(x,y,xans,yans);
+    int x3=map(xans,-0.991273,0.991273,-255,255);
+    int y3=map(yans,-0.991273,0.991273,-255,255);
 
 		/*int x2=x-512;
 		int y2=512-y;
